+++ test-app/src/main/java/org/readium/r2/testapp/reader/ReaderViewModel.kt
@@ -254,8 +254,6 @@ class ReaderViewModel(
                 else ->
                     throw IllegalStateException("Cannot create ViewModel for class ${modelClass.simpleName}.")
             }
-
-
     }
 
 }
